refactor(domain): validationSameUser() 중복 제거
@SpringBootTest를 통해 rollback 확인하여 중복 제거
