+++ src/test/java/qna/service/QnaServiceTest.java
@@ -1,5 +1,13 @@
 package qna.service;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -7,17 +15,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import qna.CannotDeleteException;
-import qna.domain.*;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import qna.domain.Answer;
+import qna.domain.DeleteHistories;
+import qna.domain.DeleteHistory;
+import qna.domain.Question;
+import qna.domain.TestQuestionFactory;
+import qna.domain.UserTest;
 import qna.repository.QuestionRepository;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class QnaServiceTest {
@@ -102,27 +106,5 @@ class QnaServiceTest {
         //then
         assertThat(question.isDeleted()).isTrue();
         verify(deleteHistoryService).saveAll(new DeleteHistories(deleteHistories));
-    }
-
-    @Test
-    void 질문자와_답변자_달라_예외_발생하면_질문_삭제여부는_거짓() {
-        //given
-        User writer = TestUserFactory.create("sanjigi");
-        User fakeWriter = TestUserFactory.create("javajigi");
-        Question question = TestQuestionFactory.create(writer);
-        Answer answer1 = TestAnswerFactory.create(writer, question);
-        Answer answer2 = TestAnswerFactory.create(fakeWriter, question);
-        questionRepository.save(question);
-
-        //when
-        when(questionRepository.findByIdAndDeletedFalse(question.getId())).thenReturn(Optional.of(question));
-        assertThat(question.isDeleted()).isFalse();
-        assertThat(question.answersCount()).isEqualTo(2);
-        assertThatThrownBy(() -> qnaService.deleteQuestion(writer, question.getId()))
-                .isInstanceOf(CannotDeleteException.class);
-
-        //then
-        assertThat(question.isDeleted()).isFalse();
-        assertThat(answer1.isDeleted()).isFalse();
     }
 }
